requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
pymongo==4.5.0

# Image Processing
pillow==10.4.0

# ML/Data Science
numpy==1.26.4
transformers==4.44.0

# Additional dependencies
python-dotenv==1.0.0

# PyTorch CPU version (optimized for Render deployment)
# Use extra-index-url to keep access to both PyPI and PyTorch repos
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
